backend/requirements.txt
certifi==2025.1.31
dnspython==1.16.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
lz4==4.4.3
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
protobuf==4.25.6
protoc-gen-openapiv2==0.0.1
pymongo==3.12.0
python-dateutil==2.9.0.post0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
certifi==2025.1.31
dnspython==1.16.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
lz4==4.4.3
numpy==2.2.2
pandas==2.2.3
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
protobuf==4.25.6
protoc-gen-openapiv2==0.0.1
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
distro==1.9.0
dnspython==1.16.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lz4==4.4.3
numpy==2.2.2
openai==1.61.0
pandas==2.2.3
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
protobuf==4.25.6
protoc-gen-openapiv2==0.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dnspython==1.16.0
fastapi==0.115.8
googleapis-common-protos==1.66.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lz4==4.4.3
numpy==2.2.2
openai==1.61.0
pandas==2.2.3
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
protobuf==4.25.6
protoc-gen-openapiv2==0.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0